Screencapture works in "standard" Chrome as well. Provided a SSL connection.
Canary is not needed for the demo to work. A SSL connection on the other hand
seems to be required (PermissionDeniedError).

I'm not sure if other demos also need a SSL connection so I did not commit
something like:

    if (window.location.protocol != "https:") {
      window.location.href = "https:" +
        window.location.href.substring(window.location.protocol.length);
    }

I was also thinking about adding:

    <a href="javascript:ssl">a SSL connection</a>

link, not sure how you feel about that though.